refactor: 직무, 부서 store 따로 파일 생성
