+++ src/services/recruitmentRequestService.js
@@ -46,31 +46,3 @@ export const createRecruitmentRequest = async (payload, options = {}) => {
         return apiResponse.data;
     }, options);
 };
-
-// 직무(포지션) 목록 조회
-export const fetchJobList = async (options = {}) => {
-    return withErrorHandling(async () => {
-        const response = await api.get(API.RECRUITMENT.JOB_LIST);
-        const apiResponse = ApiResponseDTO.fromJSON(response.data);
-
-        if (!apiResponse.success) {
-            throwCustomApiError(apiResponse.code, apiResponse.message, 400);
-        }
-
-        return apiResponse.data;
-    }, options);
-};
-
-// 부서 목록 조회
-export const fetchDepartmentList = async (options = {}) => {
-    return withErrorHandling(async () => {
-        const response = await api.get(API.RECRUITMENT.DEPARTMENT_LIST);
-        const apiResponse = ApiResponseDTO.fromJSON(response.data);
-
-        if (!apiResponse.success) {
-            throwCustomApiError(apiResponse.code, apiResponse.message, 400);
-        }
-
-        return apiResponse.data;
-    }, options);
-};
